gfs2 log recovery: Use extent list
To avoid going through the bitmap translation layer for journal writes,
GFS2 first builds up an extent list, and looks up the block mapping in
that list.  This works because journals generally consist of only few
extents (usually, just one), and the mappings stay constant while a
filesystem is mounted.

By initializing the extent list earlier, we can use it for journal
recovery as well.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>